Moved CWL-related code from lgweb to dgl.dropmake.cwl. Added dlg tool command for cwl translator (dlg cwl).
