+++ daliuge-translator/dlg/translator/tool_commands.py
@@ -284,6 +284,25 @@ def dlg_submit(parser, args):
     with _open_i(opts.pg_path) as f:
         submit(json.load(f), opts)
 
+
+def cwl(parser, args):
+    print("cwl tool")
+
+    tool.add_logging_options(parser)
+    _add_output_options(parser)
+    apps = _add_unroll_options(parser)
+    _add_partition_options(parser)
+    (opts, args) = parser.parse_args(args)
+    tool.setup_logging(opts)
+
+    # get the pgt
+    pgt = unroll(opts.lg_path, opts.oid_prefix, zerorun=opts.zerorun, app=apps[opts.app])
+    partition(pgt, opts)
+
+    # create the CWL workflow
+    from ..dropmake.cwl import create_workflow
+    create_workflow(pgt, "", "workflow.cwl", "output.zip")
+
 def register_commands():
     tool.cmdwrap('lgweb', 'A Web server for the Logical Graph Editor', 'dlg.dropmake.web.lg_web:run')
     tool.cmdwrap('submit', 'Submits a Physical Graph to a Drop Manager', dlg_submit)
@@ -292,3 +311,4 @@ def register_commands():
     tool.cmdwrap('partition', 'Divides a Physical Graph Template into N logical partitions', dlg_partition)
     tool.cmdwrap('unroll-and-partition', 'unroll + partition', dlg_unroll_and_partition)
     tool.cmdwrap('fill', 'Fill a Logical Graph with parameters', fill)
+    tool.cmdwrap('cwl', 'Translate a Logical Graph into a Common Workflow Language (CWL) workflow', cwl)
